Add file
Reads all the IP SLA data for a given date range into a Pandas dataframe.  The dataframe then calculates the start and end times of four different categories of performance.  Three levels of degradation and category for unavailable.  It then adds these ranges back into the database for easy access later.  Finally it calls a function in the DB which links together ranges that overlapped between script runs.  This script is run on a semi-regular basis typically for the purpose of finding the number and duration of WAN outages.  Not the prettiest script but one of my first attempts using pandas for data analysis.